+++ cmd/frontend/internal/app/router/router.go
@@ -21,13 +21,14 @@ const (
 
 	Logout = "logout"
 
-	SignIn            = "sign-in"
-	SignOut           = "sign-out"
-	SignUp            = "sign-up"
-	SiteInit          = "site-init"
-	VerifyEmail       = "verify-email"
-	ResetPasswordInit = "reset-password.init"
-	ResetPasswordCode = "reset-password.code"
+	SignIn             = "sign-in"
+	SignOut            = "sign-out"
+	SignUp             = "sign-up"
+	SiteInit           = "site-init"
+	VerifyEmail        = "verify-email"
+	ResetPasswordInit  = "reset-password.init"
+	ResetPasswordCode  = "reset-password.code"
+	CheckUsernameTaken = "check-username-taken"
 
 	RegistryExtensionBundle = "registry.extension.bundle"
 
@@ -72,6 +73,8 @@ func newRouter() *mux.Router {
 	base.Path("/-/sign-out").Methods("GET").Name(SignOut)
 	base.Path("/-/reset-password-init").Methods("POST").Name(ResetPasswordInit)
 	base.Path("/-/reset-password-code").Methods("POST").Name(ResetPasswordCode)
+
+	base.Path("/-/check-username-taken/{username}").Methods("GET").Name(CheckUsernameTaken)
 
 	base.Path("/-/static/extension/{RegistryExtensionReleaseFilename}").Methods("GET").Name(RegistryExtensionBundle)